requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.11.0
loguru==0.7.2
pytest==8.0.0
httpx==0.27.0
# Google Cloud APIs
google-cloud-vision==3.7.0
google-cloud-videointelligence==2.13.0
google-cloud-speech==2.34.0
google-generativeai==0.8.5
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pillow==10.2.0
aiohttp==3.9.3
